requirements.txt
insightface>=0.7.3
onnxruntime>=1.16.4
opencv-python>=4.9.0,<4.12
numpy>=1.24,<2
rich>=13.7
typer>=0.9
python-dotenv>=1.0
torch>=2.1
torchvision>=0.16
ultralytics>=8.1
deep-sort-realtime>=1.3.2
filterpy>=1.4
torchreid>=0.2.5
gdown>=5.2
tensorboard>=2.20
scikit-learn>=1.4
kagglehub>=0.2.5
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
# Database
pymongo>=4.6.0